[REFACTOR] Remove unused company type and update import paths for company data
