+++ job-match-fe/src/app/components/company/company-client.tsx
@@ -2,11 +2,10 @@
 
 import React, { useState, useEffect } from "react";
 import CompanyV1Area from "../company/company-v1-area";
-import CompanyV1Filter from "../company/filter/company-v1-filter";
 import { useCompanies } from "../company/hooks/useCompanies";
 
 const CompanyV2Client = () => {
-  const { companies } = useCompanies(); // Fetch all companies
+  const { companies } = useCompanies();
   const [searchTerm, setSearchTerm] = useState("");
   const [filteredCompanies, setFilteredCompanies] = useState(companies);
