+++ job-match-fe/src/app/components/company/company-grid-item.tsx
@@ -2,12 +2,15 @@ import React from "react";
 import Image from "next/image";
 import Link from "next/link";
 import profile_icon_1 from "@/assets/dashboard/images/icon/icon_23.svg";
-import { ICompany } from "@/types/company-type";
+import { CompanyDetails } from "@/data/company-data";
 
-const CompanyGridItem = ({ item }: { item: ICompany }) => {
+const CompanyGridItem = ({ item }: { item: CompanyDetails }) => {
   return (
     <div
-      className={`company-grid-layout ${item.isFav ? "favourite" : ""} mb-30`}
+      className={`company-grid-layout ${false ? "favourite" : ""} mb-30`}
+      style={{
+        backgroundColor: "#eff6f3",
+      }}
     >
       <Link
         href={{
@@ -16,11 +19,11 @@ const CompanyGridItem = ({ item }: { item: ICompany }) => {
         className="company-logo me-auto ms-auto rounded-circle"
       >
         <Image
-          src={item.img ? item.img : profile_icon_1}
+          src={item.logo || profile_icon_1}
           alt="image"
           className="lazy-img rounded-circle"
-          height={85}
-          width={85}
+          height={55}
+          width={55}
           style={{
             width: "100%",
             height: "100%",
@@ -39,21 +42,44 @@ const CompanyGridItem = ({ item }: { item: ICompany }) => {
           {item.name}
         </Link>
       </h5>
-      <p className="text-center mb-auto">{item.location}</p>
+      <div
+        className="team-text"
+        style={{
+          textAlign: "center",
+        }}
+      >
+        <Link
+          href={item?.website_url || `mailto:${item?.email}`}
+          className="website-btn-two tran3s"
+          target="_blank"
+        >
+          {item?.website_url || item?.email}
+        </Link>
+      </div>
+      <div
+        className="team-text"
+        style={{
+          textAlign: "center",
+          marginTop: "10px",
+          color: "#31795a",
+        }}
+      >
+        <p>{item.description.slice(0, 138)}...</p>
+      </div>
       <div className="bottom-line d-flex">
         <Link
           href={{
             pathname: `/company/${item.id}`,
           }}
         >
-          {item.vacancy} Vacancy
+          {item.active_job_ads} open jobs
         </Link>
         <Link
           href={{
             pathname: `/company/${item.id}`,
           }}
         >
-          <i className="bi bi-bookmark-dash"></i> Save
+          View
         </Link>
       </div>
     </div>
