+++ job-match-fe/src/app/components/company/company-list-item.tsx
@@ -2,13 +2,11 @@ import React from "react";
 import Image from "next/image";
 import Link from "next/link";
 import profile_icon_1 from "@/assets/dashboard/images/icon/icon_23.svg";
-import { ICompany } from "@/types/company-type";
+import { CompanyDetails } from "@/data/company-data";
 
-const CompanyListItem = ({ item }: { item: ICompany }) => {
+const CompanyListItem = ({ item }: { item: CompanyDetails }) => {
   return (
-    <div
-      className={`company-list-layout ${item.isFav ? "favourite" : ""} mb-20`}
-    >
+    <div className={`company-list-layout ${false ? "favourite" : ""} mb-20`}>
       <div className="row justify-content-between align-items-center">
         <div className="col-xl-5">
           <div className="d-flex align-items-xl-center">
@@ -19,11 +17,11 @@ const CompanyListItem = ({ item }: { item: ICompany }) => {
               className="company-logo rounded-circle"
             >
               <Image
-                src={item.img || profile_icon_1}
+                src={item.logo || profile_icon_1}
                 alt="image"
                 className="lazy-img rounded-circle"
-                height={85}
-                width={85}
+                height={55}
+                width={55}
                 style={{
                   width: "100%",
                   height: "100%",
@@ -32,7 +30,15 @@ const CompanyListItem = ({ item }: { item: ICompany }) => {
                 }}
               />
             </Link>
-            <div className="company-data">
+            <div
+              className="company-data"
+              style={{
+                display: "flex",
+                flexDirection: "column",
+                justifyContent: "center",
+                textAlign: "left",
+              }}
+            >
               <h5 className="m0">
                 <Link
                   href={{
@@ -43,31 +49,24 @@ const CompanyListItem = ({ item }: { item: ICompany }) => {
                   {item.name}
                 </Link>
               </h5>
-              <p>{item.location}</p>
+              <div className="team-text">
+                <Link
+                  href={item?.website_url || `mailto:${item?.email}`}
+                  className="website-btn-two tran3s"
+                  target="_blank"
+                >
+                  {item?.website_url || item?.email}
+                </Link>
+              </div>
             </div>
           </div>
         </div>
         <div className="col-xl-4 col-md-8">
           <div className="d-flex align-items-center ps-xxl-5 lg-mt-20">
             <div className="d-flex align-items-center">
-              {/* <Image
-                src={team_img_1}
-                alt="team_img"
-                className="lazy-img rounded-circle team-img"
-              />
-              <Image
-                src={team_img_2}
-                alt="team_img"
-                className="lazy-img rounded-circle team-img"
-              />
-              <Image
-                src={team_img_3}
-                alt="team_img"
-                className="lazy-img rounded-circle team-img"
-              /> */}
               <div className="team-text">
                 <span className="text-md fw-500 text-dark d-block">
-                  {item.successfull_matches}{" "}
+                  {item.successful_matches}{" "}
                 </span>{" "}
                 Successful Matches
               </div>
@@ -82,7 +81,7 @@ const CompanyListItem = ({ item }: { item: ICompany }) => {
               }}
               className="open-job-btn text-center fw-500 tran3s me-2"
             >
-              {item.vacancy} open jobs
+              {item.active_job_ads} open jobs
             </Link>
             <Link
               href={{
