+++ job-match-fe/src/app/components/company/filter/filter-company-location.tsx
@@ -1,12 +1,13 @@
 import React from "react";
 import NiceSelect from "@/ui/nice-select";
-import { useCompanies } from "../hooks/useCompanies";
+import { CompanyDetails } from "@/data/company-data";
 
-const FilterCompanyLocation = () => {
-  const { companies, loading } = useCompanies();
-  const company_data = companies;
-
-  const uniqueLocations = [...new Set(company_data.map((c) => c.location))];
+const FilterCompanyLocation = ({
+  companies,
+}: {
+  companies: Array<CompanyDetails>;
+}) => {
+  const uniqueLocations = [...new Set(companies.map((c) => c.city))];
   const handleLocation = (item: { value: string; label: string }) => {};
   const options = uniqueLocations.map((l) => {
     return { value: l, label: l };
@@ -15,6 +16,7 @@ const FilterCompanyLocation = () => {
     <NiceSelect
       options={options}
       defaultCurrent={0}
+      placeholder="Location"
       onChange={(item) => handleLocation(item)}
       name="Location"
     />
