+++ job-match-fe/src/app/components/company/hooks/useCompanies.tsx
@@ -1,15 +1,14 @@
 import { useState, useEffect } from "react";
-import { getCompanies } from "@/data/company-data";
-import { ICompany } from "@/types/company-type";
+import { getAllCompanies, CompanyDetails } from "@/data/company-data";
 
 export const useCompanies = () => {
-  const [companies, setCompanies] = useState<ICompany[]>([]);
+  const [companies, setCompanies] = useState<CompanyDetails[]>([]);
   const [loading, setLoading] = useState<boolean>(false);
 
   const fetchCompanies = async () => {
     setLoading(true);
     try {
-      const companies = await getCompanies();
+      const companies = await getAllCompanies();
       setCompanies(companies);
     } catch (error) {
       console.error("Error fetching companies:", error);
